requirements.txt
Flask==0.9
Jinja2==2.6
PIL==1.1.7
Werkzeug==0.9.4
glass.py==0.0.1
rauth==0.6.1
requests==1.2.3
wsgiref==0.1.2
